feat: add sub header to layout, move components
