+++ apps/trading/client-pages/market/use-market-selector-list.ts
@@ -3,7 +3,7 @@ import orderBy from 'lodash/orderBy';
 import { MarketState } from '@vegaprotocol/types';
 import { calcCandleVolume, useMarketList } from '@vegaprotocol/markets';
 import { priceChangePercentage } from '@vegaprotocol/utils';
-import type { Filter } from './market-selector';
+import type { Filter } from '../../components/market-selector/market-selector';
 import { Sort } from './sort-dropdown';
 
 // Used for sort order and filter
