+++ apps/trading/client-pages/markets/markets-page.tsx
@@ -15,7 +15,7 @@ export const MarketsPage = () => {
     updateTitle(titlefy(['Markets']));
   }, [updateTitle]);
   return (
-    <div className="h-full pt-0.5 pb-3 px-1">
+    <div className="h-full pt-0.5 pb-3 px-1.5">
       <div className="h-full my-1 border border-default rounded-sm">
         <Tabs storageKey="console-markets">
           <Tab id="all-markets" name={t('All markets')}>
