+++ apps/trading/components/header/header.tsx
@@ -12,19 +12,17 @@ export const Header = ({ title, children }: TradeMarketHeaderProps) => {
     'grid',
     'grid-rows-[min-content_min-content]',
     'xl:grid-cols-[min-content_1fr]',
-    'lg:border-x border-b border-default',
+    'border-b border-default',
     'bg-vega-clight-800 dark:bg-vega-cdark-800'
   );
   return (
-    <header className="lg:px-1">
-      <div className={headerClasses}>
-        <div className="flex flex-col justify-center items-start pl-3 lg:pl-4 pt-2 xl:pb-2 pb-0">
-          {title}
-        </div>
-        <div data-testid="header-summary" className="min-w-0">
-          <div className="px-3 lg:px-4 py-2 flex flex-nowrap gap-4 items-center text-xs overflow-x-auto">
-            {children}
-          </div>
+    <header className={headerClasses}>
+      <div className="flex flex-col justify-center items-start pl-3 lg:pl-4 pt-2 xl:pb-2 pb-0">
+        {title}
+      </div>
+      <div data-testid="header-summary" className="min-w-0">
+        <div className="px-3 lg:px-4 py-2 flex flex-nowrap gap-4 items-center text-xs overflow-x-auto">
+          {children}
         </div>
       </div>
     </header>
