+++ apps/trading/components/market-header/index.ts
@@ -1,0 +1,1 @@
+export * from './market-header';
