+++ apps/trading/components/market-selector/index.ts
@@ -1,0 +1,2 @@
+export * from './market-selector';
+export * from './market-selector-item';
