+++ apps/trading/components/market-selector/market-selector-item.spec.tsx
