+++ apps/trading/components/market-selector/market-selector.spec.tsx
